Checkin autogenerated thrift files
Summary:
This change adds the files generated by thrift compiler to the repo.
This is needed to use this package in the uber-minions repo.
The good news is that the auto generated files will not change unless
the thrift IDL file changes, even when make clean is used so it would
not create a lot of overhead for other check-ins.

Reviewers: samar, sivakk

Reviewed By: sivakk

Differential Revision: https://code.uberinternal.com/D653858