Added support for GetWorkflowExecutionHistory for completed executions
Summary:
Thrift changes to add GetWorkflowExecutionHistory api on service.

Engine changes to support GetWorkflowExecutionHistory API.  It uses
HistoryEngine to directly query the execution info for a given pair for
workflowId and runId.

CassandraPersistence layer now uses both workflowId and runId as the row
key for each executions.  This gives us the capability to not delete
execution info after the execution completes.

Each CreateExecution now creates two rows.  First one for idempotent
workflow creation based on workflowId and second one to keep state for
the actual execution.  DeleteWorkflowExecution than deleted the global
row for specific workflowId and puts a ttl on the workflow execution for
runid to be deleted after 7 days.

Reviewers: maxim, tamer, sivakk

Reviewed By: sivakk

Subscribers: aravindv, venkat

Differential Revision: https://code.uberinternal.com/D666446