History service thrift handler and client
Summary:
This revision adds the thrift handler for the history service, as well as
client interface and simple TChannel based implementation for other services
to communicate with the history service.
In the future this client will be ringpop aware and handle routing the request
to the correct shard owner, all of which is hidden from the callers.

Test Plan: TBD

Reviewers: sivakk, samar

Reviewed By: samar

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D686057