Client Factory
Summary:
This revision adds a Client Factory interface.
This can be used by the services to acquire RPC clients for other
services, while masking away the underlying transport (e.g. tchannel)
It also changes the handler initializations so that they create the engines
themselves rather than accepting it in the constructor.