Address performance issues with Transfer Queue Processor
Summary:
This change addresses the performance issues we noticed with transfer
queue processor in previous stress runs.

Transfer Queue Processor now spins off 10 task workers for processing
transfer tasks for a shard.  It also hosts an AckManager which converts
individual acks to ack levels and completes transfer tasks and moves the
AckLevel forward for the shard.

GetTransferTasks and CompleteTransferTask changes to persistence layer
to conform to the new model of retrieving and completing transfer tasks.
We no longer needs a visibility time and lock token for transfer tasks
as transfer queue processor is keeping track of the ack level and moving
the entire ack level forward for the shard.  We now use range_id for the
shard to do conditional updates to transfer queue.

Reviewers: maxim, sivakk

Reviewed By: sivakk

Subscribers: aravindv, venkat

Differential Revision: https://code.uberinternal.com/D691909