Add onebox host and use it to run integration test
Summary:
This revision adds a host that starts all cadence services in a single process.
It also wires the integration test to communicate with the one-box hosted cadence
frontend.
The integration test is now controlled by a flag. By default it won't run unless
that flag is set to true, because it takes too much time.

Test Plan: Integration test is passing

Reviewers: samar, sivakk

Reviewed By: sivakk

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D700756