Workflow directory refactor
Summary:
This Moves the contents of the workflow directory under history and matching
services as appropriate.
This change disables the matching engine unit tests temporarily, as they
depend on code internal to the history service. We must either move these
dependencies into a common location, or rewrite the tests in a different way.

Test Plan: History UT and integration passing

Reviewers: maxim, sivakk, samar

Reviewed By: sivakk, samar

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D701709