Make history client sharding aware
Summary:
This changes the history service client to be aware of the number of shards
and use the service resolver to identify which host should serve a given
request, based on the workflowID in the request.
It also makes the numberOfShards a parameter, not a constant.

Test Plan: UT, integration test

Reviewers: sivakk, maxim, samar

Reviewed By: samar

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D721011