Use time.Ticker for repeated jobs
Summary:
ShardController runs acquire shards process in a reapeated fashion.
Ticker is the right abstraction to use for any reapeated jobs with same
interval.

Transfer queue processor also runs a reapeated job to poll for transfer
task but could use a different interval.  It is ok to use a Timer in
this case but updating the code to reset the Timer instead of creating a
new one each time.

Reviewers: sivakk, tamer

Reviewed By: tamer

Subscribers: jenkins, aravindv, venkat

Differential Revision: https://code.uberinternal.com/D723274