Ignore EntityNotExistsError when completing task in TransferQueueProcessor
Summary:
This revision fixes a bug where the transfer queue processor gets stuck in
an infinite loop due to trying to complete a task that is already completed.
We saw this in a stress run where the task was completed but the operation
timed-out in cassandra, so the transfer queue processor did not know the
operation succeeded and on trying a second time, it got an EntityNotExist error.

Reviewers: sivakk, samar

Reviewed By: samar

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D725095