Logging improvements
Summary:
Create a logger with shardID tag when shard is acquired and set it on
ShardContext.

Remove reduntant API where ExecutionManager and Logger are passed to
downstream components.  Now these are exposed through ShardContext.

Create logging tags for various history service components and create
logger initialized with that tag when the component is initialized.

Reviewers: tamer

Reviewed By: tamer

Subscribers: jenkins, aravindv, venkat

Differential Revision: https://code.uberinternal.com/D730819