Change to quorum consistency level for reading tasks
Summary:
We ran into issue with stress runs where query to read tasks for
transferqueue, timerqueue, or matching engine would not receive few of
the tasks due to low consistency level for read queries.  This causes
issues as once we move the ack level forward we would never be able to
process those tasks.

Reviewers: maxim, sivakk

Reviewed By: sivakk

Subscribers: jenkins, aravindv, venkat

Differential Revision: https://code.uberinternal.com/D735252