Support for requestID to retry tasks without loosing
Summary:
History service returns an error if cassandra write fails with unknown
error and on retry the task is marked as started in history.  This
results in pretty bad experience for customers are they need for timeout
to kickin before workflow can make more progress.

Updated RecordActivityTaskStarted and RecordDecisionTaskStarted API on
history service to also take in requestId.  This requestId is written to
cassandra with the StartedEvent and it is used to figure out if the
event is marked as started by the same request on retries.

Move utility helper PrettyPrintHistory to common

Reviewers: maxim, tamer, sivakk

Reviewed By: sivakk

Subscribers: jenkins, aravindv, venkat

Differential Revision: https://code.uberinternal.com/D736142