Support for renewing RangeID for shard and life-cycle
Summary:
Persistence API changes to return ShardOwnershipLostError when
conditional update fails due to RangeID.

ShardController exposes a shardClosedCh to manage cleanup of shards when
shard ownership lost is detected.

ShardContext manages the close handling and writes to shardClosedCh when
it sees ShardOwnershipLostError.

TransferQueueProcessor calls update API even if there are no changes to
trigger Shard closure if ownership is lost.

Reviewers: sivakk, tamer, maxim

Reviewed By: tamer, maxim

Subscribers: jenkins, aravindv, venkat

Differential Revision: https://code.uberinternal.com/D736505