Shard Controller life-cycle bug fix (#1)
* Shard Controller life-cycle bug fix

We were not setting the shardID on shardContext when the shard is
acquired by a host.  This caused the shardID to be intialized to 0 for
all shards.  We used the same shardID for sending the closed event on
the shardClosedCh which caused us to always remove the shardID 0.

Updated onebox setup to have support for spinning up multiple history
hosts.  Also changed the integration tests to use 2 history hosts for
running the tests.

Also made some minor logging fixes.