History service to return ShardOwnershipLostError for clients to retry (#24)
History service API returns ShardOwnershipLostError for all API calls to
let the client know about the correct host to retry the request.

ShardController returns ShardOwnershipLostError whenever it detects
client called the host for wrong shard.  This error also has the actual
host information which can be used by the client to retry the call.

All HistoryEngine API calls checks for ShardOwnershipLostError from
persistence layer and wraps it with ShardOwnershipLostError for the
history service calls with the information for correct host to contact.