Use LocalQuorum for cassandra (#68)
Require LocalQuorum for Cassandra operations.
In a multi-DC setting, make sure the client only sees hosts from the local DC so that the request always lands on a local Cassandra node and a local quorum is guaranteed. Write could still be replicated cross-DC as well, but client should not block on its completion.