Limit the result of GetTimerTasks query (#195)
Implementation of GetTimerTasks query was sending an unbounded query to
server and then iterating over the result to only get bounded items from
it.  This ended up in generating a large amount of load on the server
where it has to return very large result set to the client as the number
of timer tasks increased in the system.
Updated the query to bound the result to limited number of tasks when
issueing query to server, instead of filtering on the client.