Apply replication history events to passive cluster (#643)
Created history replicator which is invoked by the replicator for
processing of history replication tasks.  It processes the history
events from the replication task and make mutable state updates for each
event.  Once all events are processed it commits the entire update using
the workflowContext API used by rest of the stack.

Also mutableStateBuilder changes to apply mutable state changes using
the actual event.  This required some refactoring of mutableStateBuilder
to reuse as much code as possible between replicator and rest of the
stack.

History service has new API ReplicateEvents which is called by the
replicator to apply history events.

Current this change only works for the happy case and is not guarded by
version updates on the domain.
Replicator does not support out of order processing of history
replication tasks.