Backoff on shard creation and adjust default number of persistence conns (#876)
Shard creation happens on start of each history service host.  So if we
have large number of history hosts then we could overwhelm the cluster.
This change puts in a backoff shard creation and also update the logic
to only create shards if neccessary by doing a get first.
Also adjusted the default number of persistence connection for history
and execution manger to 50.