Make task processors reselient to LimitExceededError (#891)
We have system protection in place which could result in
LimitExceededError while trying to update workflow execution.  This
protection could result in history engine restarts if transfer/timer
processing is blocked due to this error.
This change makes the processors resilient to LimitExceededError and
emits a critical log and metric and skips such tasks.
We expect an alert being setup on any critical level error and someone
manually addressing the cause of such errors rather than resulting in
complete halt of queue processing.