Bugfix for missing heartbeat timer (#1108)
Heartbeat timer can get lost if heartbeat flag is cleared on the mutable
state but never written to the database.  This could cause the heartbeat
timer flag to be still set without any actual timers in the database if
the mutable state for the execution gets reloaded.  This fix forces the
update to the database after the heartbeat timer flag is reset.