Remove cancel failed events for activity and timer
Whenever workflow worker makes a cancel decision, requestCancelActivity
in the case of an activity and cancelTimer in the case of timer we
no longer fail the individual decision if we cannot find the activity
or timer to cancel.  This means sdk code has a bug which resulted in
client code to be able to make cancellation decision in the first
place.  We should be failing the entire decision like we do in all
other places.
RequestCancelActivityTaskFailed was already removed in PR #355, so
this PR removes the unused EVENT_TYPE_REQUEST_CANCEL_ACTIVITY_TASK_FAILED
enum.
Removing all logic which results in CancelTimerFailed event in the
history when TimerCancel decision is made and instead fail the
entire decision.