Update ShardID to start with 1 instead of 0
ShardID is passed in as a field for multiple protobuf messages.  We
have created a guideline to not use 0 as a valid value for protobuf
field in a message.
Previously we used 0 as a valid ShardID which makes it impossible to
differentiate between ShardID field as not set versus caller
specifically passing in 0.
This change forces ShardID to start with 1 instead of 0 for
implementation to validate if ShardID is not passed in where
it is a required field.