Fixes MySQL persistence for workflow reset (#713)
After 2dc removal we trigger a slightly different codepath on mysql
persistence when resetting workflows.
AppendHistoryNodes when called with IsNewBranch add both
history_tree metadata and history_node.  This caused the problem
because history_node is already added when fork was called so
adding it again failed with duplicate error.  Updated
AppendHistoryNodes api to upsert instead.

Another issue was with create workflow execution never expected
to be called with `CreateWorkflowModeContinueAsNew` mode.  With
NDC now workflow reset is implemented in similar fashion as
doing continue as new.  So removed the check to disallow
`CreateWorkflowModeContinueAsNew` on create execution and added
appropriate validation.