Fixes Internal Server Error returned on simultaneous SignalWithStart requests for the same WF ID (#719)
This PR fixes a bug where simultaneous SignalWithStart requests for the same WorkflowID randomly results in one of them failing with a "Workflow Execution Already Started Error." This error does not make sense given that SignalWithStart is supposed to seamlessly handle both the case where the WF is already running and the case where it needs to be created.

Our throughput stress scenario easily reproduces this issue in the one box environment. The issue stopped reproducing once this change was integrated.

This is a relatively low-risk change. We only retry under certain error conditions that we expect to be transient, and in the worst case, we keep retrying until the request times out.