Sets ContinueAsNewInitiator value to "UNSPECIFIED" instead of "WORKFLOW" for user-created workflows (#753)
For Workflows that are explicitly started (e.g. not cron, continue-as-new, or retry-based), we were setting the ContinueAsNewInitiator type to "WORKFLOW." This can be quite confusing to the user, so we have instead switched it to "UNSPECIFIED."

In addition to updated unit tests, and running integration tests, we also inspected real-time event values for retry, cron and continue-as-new scenarios to validate proper behavior. We also scanned Temporal Server logs to make sure there was no unexpected behavior due to this change.

This is a low-risk change