Fixes instances where an activity heartbeat timeout would get "lost" (#771)
If, for any reason, the cached in-memory mutable state of a Workflow were to be reset (e.g. shard loading or transient persistent error), any in-flight activities w/heartbeats for that Workflow would lose their associated heartbeat timer tasks. In other words, it could result in an activity whose heartbeat has long expired, but which the server doesn't trying to re-schedule.

The fix is to populate the pendingActivityTimerHeartbeats map when loading the mutable state from the persistence layer if we have indicated that a heartbeat task has been created. This ensures that we don't have a pending activity without an in-memory record of a heartbeat when processing timer tasks.