Adds nil parameter to ScanCAS call() #808
When running against certain Cassandra-compatible implementations (such as ScyllaDB), the ScanCAS call needs to take in as many arguments as there are parameters that are expected to be returned. We had an incorrect number of parameters for another one of our cassandra queue methods. We fix this so that Temporal can work properly.

Tested against a real-life ScyllaDB cluster. Also unit-tests/integration-tests will fail if this breaks anything. This is a low-risk change.