Adds support for --tls_enable_host_verification as global option for TCTL (#873)
By default, tctl does not perform host verification on the temporal cluster it is talking to (if using TLS). The underlying code checks a flag, but the flag is not exposed as a Global option. This PR exposes the flag as a global option.

Validated this fix by running against a private environment with a mismatched server-name for the TLS cert. With the flag enabled, tctl gives the appropriate error message when communicating with the private environment