Makes TaskType constraint in dynamic configuration optional when specifying Task Queue Partitions #904
We've noticed that a lot of users, when attempting to override TaskQueue partition count, don't realize that they need to explicitly set the taskType and have an entry for both "Workflow" and "Activity." In most cases, users are fine just scaling both simultaneously and don't need to care about the taskType.

We've changed the dynamic configuration so that it will match a constraint if the user chooses to omit taskType as well. It first searches for a constraint with the specific taskType set, and if it cannot find it, will search for a constraint w/o taskType.

Test using both unit tests and the local one box environment.

Users might be incorrectly thinking they have task queue partitioning enabled. After upgrading to this change, they will actually have task queue partitioning enabled. They won't notice a functional difference, but may notice a performance difference.