Enable NetworkTopologyStrategy for cassandra
Add a --datacenter flag to temporal-cassandra-tool on create that will
use NetworkTopologyStrategy for replication with the specified
replication factor.

This is highly useful for clusters with topologies more complex than
either single node or single node per rack with number of racks equal to
replication factor. Any other layout will benefit from this.