Update server tag to support DB=postgresql
In 1.1.1 the start script was adjusted to expect `DB=postgresql` instead of `DB=postgres` to decide which service to configure/wait for. The compose file was updated to `postgresql` instead of `postgres` but the default tag is of the older version whose `start.sh` still expects `postgres`.

For people using this compose file without setting `SERVER_TAG` this will result in temporal hanging, mistakenly waiting for cassandra instead of postgres.