Fix enabled flag for TLS-FrontendClientConfig
The frontend client config must be considered if TLS has been configured for the
frontend and not if the internode TLS has been configured.

If it’s checked against internode it’s not possible to enable TLS only for frontend
communication. And as a result the worker service is unable to communicate
with the frontend service if TLS is enabled for the frontend only.