Sets Reconnection policy on gocql.ClusterObject (#1652)
Enabling a Reconnection policy will enable better resiliency in the event that Cassandra nodes happen to churn for a given cluster. We have seen situations where nodes churned, and the only way to recover Temporal was by initiating a reboot. This should address this.

Validated change in a deployed test environment. This is a low-risk change.