Use an SDK client to start workflows.
Using the SDK client brings the tctl code more inline with standard way
for users to start workflows and makes it easier for contributors to
adjust tctl and make use of SDK features such as interceptors/context
propagation.

This will also be useful in future if we wish to provide more plugin
APIs to control the SDK client.

This also avoids using more complex code paths for parsing memo and
search attributes given that we can only be receiving untyped JSON
inputs from the command line anyway.