Spin loading initial taskWriter state in Start()
The taskWriter isn't going to be able to write anything anyway until it
gets a task ID block so either we block in Start() or we block the first
write. This task ID block allocation happens in a background goroutine
in either case so it doesn't actually prefent Start() from returning.