Spin loading initial taskWriter state in Start() (#1842)
Spin loading initial taskWriter state

The taskWriter isn't going to be able to write anything anyway until it
gets a task ID block so either we block at startup or we block the first
write. This task ID block allocation happens in a background goroutine
in either case so it doesn't actually prefent Start() from returning.