Make taskWriter multi-stop safe
taskWriter closes a channel to tell its background goroutine(s) to stop.
This is fine and generally works but care must be taken to avoid closing
the channel twice as the second close will panic. The approach taken
here is to use a *goro.Handle for handling goroutine shutdown rather
than the channel. Multiple calls to close or Cancel() the *goro.Handle
are safe.