Fix TaskWriter panic when shutdown (#1892)
* A context-aware version of backoff.Retry

Adds backoff.RetryContext that gives the same behavior as backoff.Retry
but which also takes a context.Context parameter that allows the caller
to cancel or timeout retry execution.

* Add goro.Handle for managing goroutines

The *goro.Handle type acts as a handle for the context.Context +
context.CancelFunc + "done" channel that you need to request cancelation
of background goroutines.

* Make taskWriter multi-stop safe

taskWriter closes a channel to tell its background goroutine(s) to stop.
This is fine and generally works but care must be taken to avoid closing
the channel twice as the second close will panic. The approach taken
here is to use a *goro.Handle for handling goroutine shutdown rather
than the channel. Multiple calls to close or Cancel() the *goro.Handle
are safe.