Add incarnation ID to taskQueueManager fatal signal
A taskQueueManager shutting down can result in multiple upcalls to
unload the queue on different goroutines. We want only the first of
those calls to succeed and for later calls to have no effect. If the
later calls are allowed to run, they will unload a successor to the
taskQueueManager that emitted the signal in the first place.