Remove namespace/ dependence on cluster metadata
Removes the ClusterInfo interface for a simple bool parameter to the
cache constructure defining whether or not global namespaces are enabled
on the current cluster. Also refactored CacheEntry functions that relied
on a priori knowledge of the name of the cluster on which they are
currently being executed to taking that information as a parameter from
callers.