Adds flag to temporal-cassandra-tool to disable Initial Host Lookup #2594
Adds optional flag to disable initial host lookup when performing operations via the temporal-cassandra-tool. This is useful for some configurations of Cassandra where not having this flag enabled can significantly increase connection time or cause other issues. Validated that this worked against a test cluster.

Note that the server configuration also exposes this option, so we are just achieving parity between the server configuration and the tool configuration