Add gocql tracing
The version of gocql tracing in otel-contrib doesn't handle batch
queries very well (doesn't capture the statements) so that's a
non-starter for us. We tried to for otel-contrib under
github.com/temporalio but we need to avoid using `replace` direcives in
the server's go.mod since it's consumed as a library by several
downstream projects. All of which is to say that there's little point in
maintaining a possibly long-lived fork of otel-contrib when it's easier
just to add one file
(common/persistence/nosql/nosqlplugin/cassandra/gocql/tracing.go)
in-tree.