Honor original branch token instead of deserializing as branch info and then re-serializing
This fixes the branch token persisted in history tree info as being
potentially lossy. This change ensures forward/backward compatibility
by writing both branch info as well as the branch token. On read, the
logic fallbacks to branch info if branch token is not found.

The double write to the deprecated branch info may be removed in the
future, afterwhich no running temporal service with data may be
downgraded to "before" this change with fallback reads.

Similarly, the fallback read may be removed in the future, afterwhich
no running temporal service with data may be upgraded from "before"
this change with double writes.