Added OPA Claim mapper
OPA claim mapper simply passes through
AuthInfo via the Claim's Extensions. This
allows OPA to validate the AuthInfo as
needed.

Other AuthInfo variables, like Subject, are
left as their default values. This is to
ensure that all validation takes place
within OPA, so that temporal can be ignorant
of how validation is performed.