Removed json tags from OPA authorizer structs
Removed json tags from OPA authorizer structs
as a lot of built in structs marshal using PascalCase.
The tags were set to use camelCase for the top level
properties. This caused the rego policies to look
perculiar due to the mix of casing.

This also resulted in less files being modified
for the OPA change overall.