Add workflow.Observer to watch workflow.Context
The trickiness here is that workflow.Context instances can be paged in
and out of a cache so you don't want your observer watching a Context
that has been evicted. The work here allows observers to be reconnected
to a Context when it is loaded back into cache. The Observer module
makes not assumption about how observer registration or notification is
performed, it only promises to call the Connect function on all
appropriate Observer instances, concrete implementations of which will
do the type-specific registration and notification.