Add automaxprocs.
This will automate the setting of GOMAXPROCS in Kubernetes/docker environments
where it is not already set as part of the deployment.

Setting GOMAXPROCS to match resource limits (rather than the total core count
of the node) allows Go to more efficiently use the available cores and
reduces CPU throttling (eliminating it if limits are set to an integer number
of cores).

This issue was highlighted during benchmarking but probably effects a large
number of real world Kubernetes deployments where CPU limits are set but
GOMAXPROCS is not.