Add early delivery of messages for CompleteWFE cmd
Closing a workflow with incomplete updates causes an error but it may be
the case that the messages that complete the incomplete updates are in
the same WFTCompleted message as the CompleteWorkflowExecution command.
In such cases becuase we currently deliver all messages _after_ all
commands, the fact that the update completed is not known at the time we
attempt to handle the CompleteWorkflowExecution command and thus the
WFTCompletion fails. Here we add (temporarily) the ability to
early-deliver messages for this one command type. Soon the messages will
be sequenced in with the commands so that this problem goes away.