Use DefaultServices for server_test (#4566)
<!-- Describe what has changed in this PR -->
**What changed?**
1. We now use `ForServices(DefaultServices)` in the server test. 
2. In addition, we only check error messages while the server is
starting and running, not while it is shutting down.

<!-- Tell your future self why have you made these changes -->
**Why?**
1. This test wasn't doing much without ForServices, which I mistakenly
removed in a previous PR. It was just starting the server but no
services.
2. The volume and variety of errors during shutdown is intractable

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
I ran it locally 10 times, and twice on the CI.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Could easily become a flaky test.

<!-- Is this PR a hotfix candidate or require that a notification be
sent to the broader community? (Yes/No) -->
**Is hotfix candidate?**
No