add stop reason for context (#4601)
<!-- Describe what has changed in this PR -->
**What changed?**
When a shard context is stopped, it stores whether or not that was due
to a shard ownership lost (SOL) error.

<!-- Tell your future self why have you made these changes -->
**Why?**
Upcoming changes to the shard controller will delay re-acquiring a shard
that recently had an SOL error.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Tested locally & ci, and in a staging environment (with other changes).

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
None; this change is only used by tests.

<!-- Is this PR a hotfix candidate or require that a notification be
sent to the broader community? (Yes/No) -->
**Is hotfix candidate?**
No.